Working for large datasets.
The problem still exists where results are simply incorrect if the block-size is not equal to the tile-size.
We notice that in a small dataset, the result at 0th index is 18 (correct is 11), when at 32nd index it is 0 (correct is 7), which means somehow the 32nd index is accumulating at 0th index (the difference aligns perfectly for all places with errors).

Known problem:
Even with block-size == tile-size, it seems like we have to drop the floating point accuracy to 1e-2 for the result to equal the CPU results. This may be a floating point issue, but also could be related to something else that I am missing.